Setup style directory and started layout
I organized the style directory for development, and then started
on the layout styling. It's very basic, but some stuff is placed
properly like the navbar and the page content (top spacing).

I'm thinking each visual section should have it's own ID and be styled
accordingly. If we switch up the layout from V02 then this might change,
but to accommodate the free-floating circles, this is probably the best
route.

I don't wanna do too much more specific layout stuff, but I can add
the fonts and colors as sass variables.